Implemented a dark theme for the jetzt speedreader
Changes:
 - added a new case to handleKeydown to listen for a button press on `0`. This
   is arbitrary and subject to change.
 - added a control method toggleTheme, which does nothing other than call
   the new toggleTheme method in Reader (see next bullet).
 - added a method toggleTheme in Reader, which uses two added CSS classes,
   dark-theme and sr-pivot-dark, to change the appearance of the jetzt reader
   from black/red on white/gray to white/green on black/gray.
 - exposed wordBox as its own element in Reader, to allow for the easier
   maniuplation of its CSS classes.
 - added a variable called theme to reader, which keeps track of the current
   theme of the reader (true = light, false = dark)
 - added various comments in places where I made changes.

I saw this on the todo list and thought it was low hanging fruit that I could
tackle. This is my first contribution to an open source project. Please be
kind, and if I did royally screw something up, I would love to hear about how
I can avoid that in the future in a constructive way.

ianzapolsky